// src/domains/service-clients/course/proto/generated/course_service.ts
export interface CreateCourseRequest {
  title: string;
  topics: string[];
  instructorId: string;
  instructor: Instructor | undefined;
  subTitle: string;
  category: string;
  subCategory: string;
  language: string;
  level: string;
  subtitleLanguage: string;
  durationValue: string;
  durationUnit: string;
}

export interface GetCourseRequest {
  courseId: string;
}

export interface GetCoursesByIdsRequest {
  courseIds: string[];
}

export interface GetCoursesByIdsResponse {
  success?: CoursesResponse | undefined;
  error?: Error | undefined;
}

export interface GetCourseBySlugRequest {
  slug: string;
}

export interface GetAllCoursesRequest {
  pagination: Pagination | undefined;
}

export interface UpdateCourseRequest {
  title?: string | undefined;
  topics: string[];
  subTitle?: string | undefined;
  category?: string | undefined;
  subCategory?: string | undefined;
  language?: string | undefined;
  level?: string | undefined;
  subtitleLanguage?: string | undefined;
  durationValue?: string | undefined;
  durationUnit?: string | undefined;
  courseId: string;
  userId: string;
  description?: string | undefined;
  learningOutcomes: string[];
  targetAudience: string[];
  requirements: string[];
  thumbnail?: string | undefined;
  trailer?: string | undefined;
  status?: string | undefined;
  price?: number | undefined;
  discountPrice?: number | undefined;
  currency?: string | undefined;
}

export interface DeleteCourseRequest {
  courseId: string;
  userId: string;
}

export interface GetCoursesByInstructorRequest {
  instructorId: string;
  pagination: Pagination | undefined;
}

export interface GetEnrolledCoursesRequest {
  userId: string;
  pagination: Pagination | undefined;
}
// ...
export interface CourseResponse {
  course?: CourseData | undefined;
  error?: Error | undefined;
}

export interface CoursesData {
  courses: CourseData[];
  total: number;
}

export interface CoursesResponse {
  courses?: CoursesData | undefined;
  error?: Error | undefined;
}

export interface DeleteCourseResponse {
  success?: DeleteSuccess | undefined;
  error?: Error | undefined;
}

export interface DeleteSuccess {
  deleted: boolean;
}

/** Section Messages */
export interface CreateSectionRequest {
  courseId: string;
  userId: string;
  title: string;
  description?: string | undefined;
  order: number;
  isPublished: boolean;
}

export interface GetSectionRequest {
  sectionId: string;
}
// ...
export interface SectionResponse {
  section?: SectionData | undefined;
  error?: Error | undefined;
}

export interface SectionsData {
  sections: SectionData[];
}

export interface SectionsResponse {
  sections?: SectionsData | undefined;
  error?: Error | undefined;
}

export interface DeleteSectionResponse {
  success?: DeleteSuccess | undefined;
  error?: Error | undefined;
}

/** Lesson Messages */
export interface CreateLessonRequest {
  sectionId: string;
  courseId: string;
  userId: string;
  isPreview?: boolean | undefined;
  description?: string | undefined;
  estimatedDuration?: number | undefined;
  order?: number | undefined;
  title?: string | undefined;
  isPublished?: boolean | undefined;
  contentType?: string | undefined;
  contentUrl?: string | undefined;
  metadata: ContentMetaData | undefined;
}

export interface GetLessonRequest {
  lessonId: string;
  courseId: string;
}

export interface UpdateLessonRequest {
  lessonId: string;
  courseId: string;
  userId: string;
  sectionId: string;
  isPreview?: boolean | undefined;
  description?: string | undefined;
  estimatedDuration?: number | undefined;
  order?: number | undefined;
  title?: string | undefined;
  isPublished?: boolean | undefined;
  contentType?: string | undefined;
  contentUrl?: string | undefined;
  metadata: ContentMetaData | undefined;
}

export interface ContentMetaData {
  title?: string | undefined;
  fileName?: string | undefined;
  mimeType?: string | undefined;
  fileSize?: string | undefined;
  url?: string | undefined;
}

export interface DeleteLessonRequest {
  lessonId: string;
  courseId: string;
  userId: string;
}

export interface GetLessonsBySectionRequest {
  sectionId: string;
}

export interface LessonData {
  id: string;
  sectionId: string;
  isPreview: boolean;
  description: string;
  estimatedDuration: number;
  order: number;
  title: string;
  isPublished: boolean;
  contentType: string;
  contentUrl: string;
  metadata: ContentMetaData | undefined;
  createdAt: string;
  updatedAt: string;
  deletedAt?: string | undefined;
}

export interface LessonResponse {
  lesson?: LessonData | undefined;
  error?: Error | undefined;
}

export interface LessonsData {
  lessons: LessonData[];
}

export interface LessonsResponse {
  lessons?: LessonsData | undefined;
  error?: Error | undefined;
}
// ...
export const GetCourseRequest: MessageFns<GetCourseRequest> = {
  encode(message: GetCourseRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.courseId !== "") {
      writer.uint32(10).string(message.courseId);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetCourseRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetCourseRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.courseId = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetCourseRequest {
    return { courseId: isSet(object.courseId) ? globalThis.String(object.courseId) : "" };
  },

  toJSON(message: GetCourseRequest): unknown {
    const obj: any = {};
    if (message.courseId !== "") {
      obj.courseId = message.courseId;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetCourseRequest>, I>>(base?: I): GetCourseRequest {
    return GetCourseRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetCourseRequest>, I>>(object: I): GetCourseRequest {
    const message = createBaseGetCourseRequest();
    message.courseId = object.courseId ?? "";
    return message;
  },
};
// ...
export const GetAllCoursesRequest: MessageFns<GetAllCoursesRequest> = {
  encode(message: GetAllCoursesRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.pagination !== undefined) {
      Pagination.encode(message.pagination, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): GetAllCoursesRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseGetAllCoursesRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.pagination = Pagination.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): GetAllCoursesRequest {
    return { pagination: isSet(object.pagination) ? Pagination.fromJSON(object.pagination) : undefined };
  },

  toJSON(message: GetAllCoursesRequest): unknown {
    const obj: any = {};
    if (message.pagination !== undefined) {
      obj.pagination = Pagination.toJSON(message.pagination);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<GetAllCoursesRequest>, I>>(base?: I): GetAllCoursesRequest {
    return GetAllCoursesRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<GetAllCoursesRequest>, I>>(object: I): GetAllCoursesRequest {
    const message = createBaseGetAllCoursesRequest();
    message.pagination = (object.pagination !== undefined && object.pagination !== null)
      ? Pagination.fromPartial(object.pagination)
      : undefined;
    return message;
  },
};
// ...
export const CourseResponse: MessageFns<CourseResponse> = {
  encode(message: CourseResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.course !== undefined) {
      CourseData.encode(message.course, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CourseResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCourseResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.course = CourseData.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CourseResponse {
    return {
      course: isSet(object.course) ? CourseData.fromJSON(object.course) : undefined,
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
    };
  },

  toJSON(message: CourseResponse): unknown {
    const obj: any = {};
    if (message.course !== undefined) {
      obj.course = CourseData.toJSON(message.course);
    }
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CourseResponse>, I>>(base?: I): CourseResponse {
    return CourseResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CourseResponse>, I>>(object: I): CourseResponse {
    const message = createBaseCourseResponse();
    message.course = (object.course !== undefined && object.course !== null)
      ? CourseData.fromPartial(object.course)
      : undefined;
    message.error = (object.error !== undefined && object.error !== null) ? Error.fromPartial(object.error) : undefined;
    return message;
  },
};
// ...
export const CoursesResponse: MessageFns<CoursesResponse> = {
  encode(message: CoursesResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.courses !== undefined) {
      CoursesData.encode(message.courses, writer.uint32(10).fork()).join();
    }
    if (message.error !== undefined) {
      Error.encode(message.error, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CoursesResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCoursesResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.courses = CoursesData.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.error = Error.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): CoursesResponse {
    return {
      courses: isSet(object.courses) ? CoursesData.fromJSON(object.courses) : undefined,
      error: isSet(object.error) ? Error.fromJSON(object.error) : undefined,
    };
  },

  toJSON(message: CoursesResponse): unknown {
    const obj: any = {};
    if (message.courses !== undefined) {
      obj.courses = CoursesData.toJSON(message.courses);
    }
    if (message.error !== undefined) {
      obj.error = Error.toJSON(message.error);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<CoursesResponse>, I>>(base?: I): CoursesResponse {
    return CoursesResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CoursesResponse>, I>>(object: I): CoursesResponse {
    const message = createBaseCoursesResponse();
    message.courses = (object.courses !== undefined && object.courses !== null)
      ? CoursesData.fromPartial(object.courses)
      : undefined;
    message.error = (object.error !== undefined && object.error !== null) ? Error.fromPartial(object.error) : undefined;
    return message;
  },
};
// ...
export const CourseServiceClient = makeGenericClientConstructor(
  CourseServiceService,
  "course.CourseService",
) as unknown as {
  new (address: string, credentials: ChannelCredentials, options?: Partial<ClientOptions>): CourseServiceClient;
  service: typeof CourseServiceService;
  serviceName: string;
};